pyproject.toml
[tool.poetry]
name = "finandata"
version = "0.1.0"
description = ""
authors = ["Hailey"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
loguru = "^0.7.2"
tqdm = "^4.66.2"
pydantic = "^2.6.2"
requests = "^2.31.0"
pandas = "^2.2.1"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
APScheduler = "^3.10.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
